Fix indenting CData
The reason of all CData indenting issues was replacing CDATA
tags before indent and then replacing it back. Due to this
CDATA wasn't hanlded by minidom parser correctly.

After this fix, CDATA will remain untouched. However,
indenting CDATA block with writexml() (which is called from
toprettyxml) works wierd - it starts cdata on the beginning
of the next line (without indent) and keeps parent closing
tag on the same line as well. E.g. formatted XML might look
like:

<a>
    <content>
    <![CDATA[hello & goodbye]]> </content>
</a>

To fix this, plugin now does monkey patching of
CDataSection.writexml(). This is just a hack and potentially
destructive action, so warning message is displayed when
patched version is used and this behaviour can be disabled
in settings.

This commit fixes #57, #93, #101